Module cache_sim::condition

Implementations of conditions for filtering the frequency histogram

Structs

LastNItems

A general condition for checking the previous N items. Can take any vector of items and filters for accesses in the trace where the last N items were the exact contents of the vector.

LastNStrides

A general condition for checking the previous N strides. Can take any vector of strides (i64) and filters for accesses in the trace where the last N strides were the exact contents of the vector.

NoCondition

No condition on the trace, includes all elements for a full frequency histogram

Traits

Condition

An abstract representation of a condition on a frequency histogram.