Trait cache_sim::condition::Condition

pub trait Condition<I: Item> {
    // Required method
    fn check(&self, trace: &Trace<I>, index: usize) -> bool;
}

An abstract representation of a condition on a frequency histogram.

Required Methods

fn check(&self, trace: &Trace<I>, index: usize) -> bool

Check whether a certain element should be counted by the histogram. Needs the trace and the index of the element to check.