Struct cache_sim::replacement_policy::Landlord

pub struct Landlord<I: Item = GeneralModelItem, T: Tiebreaker<I> = Lru<GeneralModelItem>> { /* private fields */ }

The landlord replacement algotihm.

Detailed in this paper: https://arxiv.org/abs/cs/0205033

The tiebreaker (for evicting multiple zero-credit items) defaults to Lru.

use cache_sim::{Cache, Landlord, GeneralModelGenerator};

let mut cache = Cache::<Landlord, (), _>::new(3);
let mut g = GeneralModelGenerator::new();

let a = g.item(1.0, 1);
let b = g.item(0.5, 2);
let c = g.item(100.0, 2);
let d = g.item(1.0, 1);

cache.access(a);
cache.access(b);
cache.access(c);
cache.access(d);

assert_eq!(cache.set(), &HashSet::from([c, d]));

Implementations

impl<I: Item, T: Tiebreaker<I> + Default> Landlord<I, T>

pub fn new(credit_increase: f64) -> Self

Instantiate a new landlord replacement policy.

The credit_increase parameter represents the percentage of the gap between the current credit and maximum credit (cost) to increase an item’s credit when it is hit. It should not be above one. Higher values are closer to LRU, lower values are closer to FIFO. This defaults to 1, and should generally be between 0 and 1.

impl<I: Item, T: Tiebreaker<I>> Landlord<I, T>

pub fn with_tiebreaker(tiebreaker: T, credit_increase: f64) -> Self

Instantiate a new landlord replacement policy, with a specifically configured tiebreaker.

The credit_increase parameter represents the percentage of the gap between the current credit and maximum credit (cost) to increase an item’s credit when it is hit. It should not be above one. Higher values are closer to LRU, lower values are closer to FIFO. This defaults to 1, and should generally be between 0 and 1.