Trait cache_sim::stats::Stat

pub trait Stat<I: Item>: Default {
    // Required method
    fn update(&mut self, set: &HashSet<I>, next: I, to_be_evicted: &HashSet<I>);
}

An abstract representation of a cache statistic.

Importantly, this is implemented for any tuple of statistics, up to size 12 (for technical reasons we can’t go higher than this right now). So to track multiple statistics, you can use a tuple of statistics.

Required Methods

fn update(&mut self, set: &HashSet<I>, next: I, to_be_evicted: &HashSet<I>)

Update the stat.

We get passed the set before evictions occur.