Struct cache_sim::trace::Trace

pub struct Trace<I: Item = u32> { /* private fields */ }

A trace.

Implementations

impl<I: Item> Trace<I>

pub fn frequency_histogram( &self, condition: &impl Condition<I> ) -> HashMap<I, u32>

Calculate the frequency histogram of the items based on a given condition.

let frequencies = Trace::from(vec![0, 0, 1, 0, 3, 1]).frequency_histogram(&NoCondition);
assert_eq!(frequencies.get(&0), Some(&3));
assert_eq!(frequencies.get(&1), Some(&2));
assert_eq!(frequencies.get(&2), None);

pub fn stride_histogram( &self, condition: &impl Condition<I> ) -> HashMap<i64, u32>

Calculate the frequency histogram of the strides based on a given condition.

let strides = Trace::from(vec![0, 0, 4, 3, 2, 2, 0]).stride_histogram(&NoCondition);
assert_eq!(frequencies.get(&0), Some(&2));
assert_eq!(frequencies.get(&-1), Some(&2));
assert_eq!(frequencies.get(&3), None);

pub fn frequency_histogram_many<'a>( &self, conditions: &'a HashMap<String, (Box<dyn Condition<I>>, bool)> ) -> (HashMap<&'a str, HashMap<I, u32>>, HashMap<&'a str, HashMap<i64, u32>>)

Calculate several frequency histograms, each identified by a string. Can produce a mix of item and stride frequency histograms, determined by the bool accompanying the condition. (true would indicate a stride frequency histogram)

let mut conditions: HashMap<String, (Box<dyn Condition<GeneralModelItem>>,bool)> = HashMap::with_capacity(2);
conditions.insert(String::from("Items"), (Box::new(NoCondition),false));
conditions.insert(String::from("Strides"), (Box::new(NoCondition),true));
let histograms = Trace::from(vec![0, 0, 4, 3, 2, 2, 0]).frequency_histogram_many(&conditions);
assert_eq!(histograms.get(&"Items").get(&0), Some(&3));
assert_eq!(frequencies.get(&"Strides").get(&0), Some(&1));

pub fn stack_distances(&self, paging_model: bool) -> StackDistance

Calculate the stack distances. Under the paging model, size is considered 1 for all items.

use cache_sim::Trace;

let distances = Trace::from(vec![0, 0, 1, 0, 3, 0, 1]).stack_distances();
assert_eq!(
    distances.inner(),
    &[None, Some(0), None, Some(1), None, Some(1), Some(2)]
);

For more details, see StackDistance.

pub fn write_conditional_frequencies<W: Write>( &self, conditions: HashMap<String, (Box<dyn Condition<I>>, bool)>, writer: impl Fn() -> Result<W> ) -> Result<()>

Write the conditional frequencies for each condition to the output stream.

Writer is a function that can give us a writer; ideally it should return a handle to the same underlying output stream each time.

Errors

If writing to the csv fails.

TODO: figure out a non-boxed return type

pub fn average_entropy(&self, prefix: usize) -> f64

Calculates the conditional entropy of an item, conditioned on the last prefix items seen.

Each sequence of items has a distribution of the items that follow it, and this is the weighted sum of all of those.

pub fn stride_entropy(&self, prefix: usize) -> f64

Calculates the conditional entropy of a stride length, conditioned on the last prefix strides seen.

Each sequence of items has a distribution of the strides that follow it, and this is the weighted sum of all of those.

(This is analagous to average_entropy for items)

pub fn iter(&self) -> Iter<'_, I>

Get an iterator over the inner vector of items

pub fn inner(&self) -> &[I]

Get a reference to the inner vector of items.

pub fn into_inner(self) -> Vec<I>

Take ownership of the inner vector of items.

The ith element of the vector is the ith access of the trace.

pub fn strides(&self) -> &[i64]

Get a reference to the inner vector of strides.

pub fn len(&self) -> usize

Get the length of the trace.

pub fn num_items(&self, length: usize) -> usize

Get the number of unique sequences of items in the trace that are exactly length long

pub fn num_strides(&self, length: usize) -> usize

Get the number of unique sequences of strides in the trace that are exactly length long

pub fn is_empty(&self) -> bool

Check whether the trace is empty.

impl Trace<u32>

pub fn pretty_print(&self) -> String

If the elements in the trace are all smaller than 26, display them as letters instead.

let trace = Trace::from(vec![0, 0, 2, 3, 1, 14]);
assert_eq!(&trace.pretty_print(), "A, A, C, D, B, O");

Note that this doesn’t work for higher values of the item:

let trace = Trace::from(vec![1, 2, 26]);
assert_eq!(&trace.pretty_print(), "1, 2, 26");